--- a/lib/types/uuid.js
+++ b/lib/types/uuid.js
@@ -2,7 +2,6 @@
 
 const crypto = require("crypto");
 const utils = require("../utils");
-const rust = require("../../index");
 
 /** @module types */
 
@@ -19,10 +18,10 @@ class Uuid {
         /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
 
     /**
-     * @type {rust.UuidWrapper}
+     * @type {Buffer}
      * @private
      */
-    #internal;
+    #raw;
 
     /**
      * Creates a new instance of Uuid based on a Buffer
@@ -34,7 +33,7 @@ class Uuid {
                 "You must provide a buffer containing 16 bytes",
             );
         }
-        this.#internal = rust.UuidWrapper.new(buffer);
+        this.#raw = buffer;
     }
 
     /**
@@ -43,7 +42,7 @@ class Uuid {
      * @type Buffer
      */
     get buffer() {
-        return this.#internal.getBuffer();
+        return this.#raw;
     }
 
     set buffer(_) {
@@ -94,7 +93,7 @@ class Uuid {
      * @returns {Buffer}
      */
     getBuffer() {
-        return this.#internal.getBuffer();
+        return this.buffer;
     }
 
     /**
@@ -150,7 +149,7 @@ class Uuid {
      * @returns {rust.UuidWrapper}
      */
     getInternal() {
-        return this.#internal;
+        return this.#raw;
     }
 }