--- a/src/requests/parameter_wrappers.rs
+++ b/src/requests/parameter_wrappers.rs
@@ -11,7 +11,6 @@ use crate::{
         local_date::LocalDateWrapper,
         local_time::LocalTimeWrapper,
         type_wrappers::{ComplexType, CqlType},
-        uuid::UuidWrapper,
     },
     utils::{bigint_to_i64, js_error},
 };
@@ -148,10 +147,18 @@ fn cql_value_from_napi_value(typ: &ComplexType, elem: &Array, pos: u32) -> napi:
         ),
         CqlType::Time => CqlValue::Time(get_element!(&LocalTimeWrapper).get_cql_time()),
         CqlType::Timeuuid => CqlValue::Timeuuid(CqlTimeuuid::from_bytes(
-            get_element!(&UuidWrapper).get_cql_uuid().into_bytes(),
+            get_element!(Buffer)
+                .to_vec()
+                .try_into()
+                .map_err(|_| js_error("Value must be Buffer with 16 elements to be TimeUuid"))?,
         )),
         CqlType::Tuple => CqlValue::Tuple(cql_value_vec_from_tuple(typ, &get_element!(Array))?),
-        CqlType::Uuid => CqlValue::Uuid(get_element!(&UuidWrapper).get_cql_uuid()),
+        CqlType::Uuid => CqlValue::Uuid(
+            get_element!(Buffer)
+                .to_vec()
+                .try_into()
+                .map_err(|_| js_error("Value must be Buffer with 16 elements to be Uuid"))?,
+        ),
         CqlType::Varint => todo!(),
         CqlType::Unprovided => return Err(js_error("Expected type information for the value")),
         CqlType::Empty => unreachable!("Should not receive Empty type here."),